requirements.txt
torchvision~=0.16.1
torch~=2.1.1
Pillow~=10.1.0
PyQt5~=5.15.10